Check a reference chapter was supplied; die if not.
Getting the reference chapter has a quirk right now.  If a line is simply

chapter=0

the regex will not recognize the value (this also is the case if there is
white space before the zero).  The value needs to be set to "00" to be
recognized, probably because it is now being treated like a string, not a
number.