Use a unique constraint to look for references.
Use the unique constraint on (referencing_work_id, type_id, chapter, rank) to
determine if a reference already exists (to update) or does not (to create).
Thus, if a reference's text changes, the row will be updated, instead of a new
reference being inserted.

This does not handle the case where a reference's chapter, for example,
changes.  In this case, the updated version will be treated like a new
reference and will be inserted as a new record.