Update the modified timestamp with the reference.
When a reference record is created or updated, set the modified timestamp to
be the current time.  As a side effect, this will cause the modified time to
be updated every time the script runs on a reference, as this value will be
different from what is in the database every time.